requirements.txt
# Medical Imaging
pydicom>=2.4.0
nibabel>=5.0.0
SimpleITK>=2.3.0
nilearn>=0.10.0

# Deep Learning
torch>=2.0.0
monai>=1.3.0
torchvision>=0.15.0

# Image Registration
antspyx>=0.4.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
h5py>=3.8.0
pyyaml>=6.0